Update student page
