--- packages/inside/src/app/main/club/student/certificate/page.tsx
@@ -1,7 +1,0 @@
-import {CertificatePage} from "@bitgouel/common/src/pages"
-
-const Certificate = () => {
-    return <CertificatePage />
-};
-
-export default Certificate;
