+++ shared/common/src/pages/club/student/style.ts
@@ -1,4 +1,5 @@
 import styled from '@emotion/styled'
+import { unstable_getServerProps } from 'next/dist/build/templates/pages'
 
 export const SlideBg = styled.div<{ url: any }>`
   height: 15rem;
@@ -49,4 +50,165 @@ export const ClubButton = styled.div`
   &:hover {
     background-color: rgb(255, 255, 255, 0.4);
   }
+`
+
+export const CertificateWrapper = styled.div`
+  width: 100%;
+  height: 100%;
+  display: flex;
+  justify-content: center;
+`
+
+export const CertificateContainer = styled.div`
+  display: flex;
+  flex-direction: column;
+  align-items: flex-start;
+  width: 75rem;
+  height: 100%;
+  justify-content: center;
+  margin-top: 4rem;
+  gap: 2.6rem;
+`
+
+export const ProfileBox = styled.div`
+  display: flex;
+  flex-direction: column;
+  align-items: flex-start;
+  gap: 0.3rem;
+
+  h3 {
+    ${({ theme }) => theme.typo.text_lg.semibold};
+    margin: 0;
+    margin-left: 1rem;
+  }
+`
+
+export const ProfileInfoBox = styled.div`
+  display: flex;
+  flex-direction: column;
+  align-items: flex-start;
+  ${({ theme }) => theme.typo.text_md.regular};
+  color: ${({ theme }) => theme.color.gray['400']};
+  margin-left: 1rem;
+
+  span {
+    margin: 0;
+  }
+
+  b {
+    color: ${({ theme }) => theme.color.main};
+    ${({ theme }) => theme.typo.text_md.regular}
+  }
+`
+
+export const CertificateBox = styled.div`
+  display: flex;
+  flex-direction: column;
+  align-items: flex-start;
+  gap: 0.5rem;
+`
+
+export const CertificatePlusBox = styled.div`
+  display: flex;
+  align-items: center;
+  gap: 0.2rem;
+  color: ${({ theme }) => theme.color.black};
+
+  h3 {
+    ${({ theme }) => theme.typo.text_lg.semibold};
+    margin: 0;
+    margin-left: 1rem;
+  }
+
+  svg {
+    cursor: pointer;
+  }
+`
+
+export const CertificateListBox = styled.ul`
+  display: flex;
+  flex-direction: column;
+  align-items: flex-start;
+  gap: 0.4rem;
+  margin: 0;
+  padding: 0;
+  list-style: none;
+  margin-left: 1rem;
+`
+
+export const CertificateItemBox = styled.div`
+  display: flex;
+  align-items: center;
+  gap: 0.4rem;
+  ${({ theme }) => theme.typo.text_md.regular};
+
+  span {
+    &:nth-of-type(1) {
+      color: ${({ theme }) => theme.color.black};
+    }
+    
+    &:nth-of-type(2) {
+      color: ${({ theme }) => theme.color.gray['400']};
+    }
+  }
+  `
+  export const ListToggle = styled.div<{list: '추가' | "추가됨"}>`
+    width: 0.375rem;
+    height: 0.375rem;
+    background-color: ${({ theme, list }) => list === '추가' ? theme.color.gray['700'] : theme.color.main};
+    border-radius: 50%;
+  `
+
+export const CertificatePlusInputBox = styled.div`
+  display: flex;
+  align-items: center;
+  gap: 0.3rem;
+`
+
+export const AddCertificateBox = styled.div`
+  display: flex;
+  align-items: center;
+  gap: 0.4rem;
+`
+
+
+export const AddCertificateInput = styled.input`
+  border: none;
+  outline: none;
+  border-bottom: 0.0625rem solid ${({ theme }) => theme.color.main};
+  width: 12.5rem;
+  height: 1.5625rem;
+  padding: 0;
+  ${({theme}) => theme.typo.text_md.regular};
+`
+
+export const AddCertificateDateBox = styled.div`
+  display: flex;
+  align-items: center;
+  border-bottom: 0.0625rem solid ${({ theme }) => theme.color.main};
+  width: 10rem;
+  height: 1.5625rem;
+  gap: 0.6rem;
+`
+
+export const SelectDateContainer = styled.div`
+  position: relative;
+  display: flex;
+  flex-direction: column;
+  align-items: flex-start;
+
+  svg {
+    cursor: pointer;
+  }
+`
+
+export const ShowDateText = styled.span`
+  ${({theme}) => theme.typo.text_md.regular};
+  margin-bottom: 0.4rem;
+`
+
+export const AddCertificateIcon = styled.div`
+  display: flex;
+  justify-content: center;
+  align-items: center;
 `
